+++ api/app/services/recipeService.ts
@@ -1,14 +1,34 @@
 import Recipe from "#models/recipe";
 import { CreateRecipeDTO, UpdateRecipeDTO } from "../dto/recipeDTO.js";
-import { UnitEnum } from "../enums/unitEnum.js";
+import { CreateRecipeIngredientDTO, RecipeIngredientDTO, UpdateRecipeIngredientDTO } from "../dto/recipeIngredientDTO.js";
 
 export default class RecipeService {
     public async list() {
-        return Recipe.all()
+        const recipes = await Recipe.query()
+
+        return recipes.map((recipe) => ({
+            id: recipe.id,
+            name: recipe.name
+        }))
     }
 
     public async findById(id: number) {
-        return Recipe.findOrFail(id)
+        const recipe = await Recipe
+            .query()
+            .where('id', id)
+            .preload('ingredients')
+            .firstOrFail()
+        
+        return {
+            id: recipe.id,
+            name: recipe.name,
+            ingredients: recipe.ingredients.map((ingredient) => ({
+                id: ingredient.id,
+                name: ingredient.name,
+                quantity: ingredient.$extras.quantity,
+                unit: ingredient.$extras.unit
+            }))
+        }
     }
 
     public async create(data: CreateRecipeDTO) {
@@ -27,34 +47,91 @@ export default class RecipeService {
 
     public async delete(id: number) {
         const recipe = await Recipe.findOrFail(id)
+
+        if (!recipe) return false
+
         await recipe.delete()
+        return true
     }
 
-    public async addIngredient(recipeId: number, ingredientId: number, quantity: number, unit: UnitEnum) {
+    public async addIngredient(recipeId: number, data: CreateRecipeIngredientDTO) {
         const recipe = await Recipe.findOrFail(recipeId)
 
         await recipe.related('ingredients').attach({
-            [ingredientId]: { quantity, unit }
+            [data.ingredientId]: {
+                quantity: data.quantity,
+                unit: data.unit
+            }
         })
     }
 
-    public async updateIngredient(recipeId: number, ingredientId: number, quantity: number, unit: UnitEnum) {
+    public async updateIngredient(recipeId: number, ingredientId: number, data: UpdateRecipeIngredientDTO) {
         const recipe = await Recipe.findOrFail(recipeId)
 
-        await recipe.related('ingredients').sync({
-            [ingredientId]: { quantity, unit }
-        })
+        const affectedRows = await recipe
+            .related('ingredients')
+            .pivotQuery()
+            .where('ingredients.id', ingredientId)
+            .update({
+                quantiy: data.quantity,
+                unit: data.unit
+            })
+        
+        if (affectedRows.length === 0) return null
+
+        const ingredient = await recipe.related('ingredients').query().where('ingredients.id', ingredientId).first()
+        return ingredient
     }
 
     public async removeIngredient(recipeId: number, ingredientId: number) {
         const recipe = await Recipe.findOrFail(recipeId)
 
+        const ingredient = await recipe
+            .related('ingredients')
+            .query()
+            .where('ingredients.id', ingredientId)
+            .first()
+
+        if (!ingredient) return false
+
         await recipe.related('ingredients').detach([ingredientId])
+        return true
     }
 
-    public async listIngredients(recipeId: number) {
-        const recipe = await Recipe.query().where('id', recipeId).preload('ingredients').firstOrFail()
+    public async indexIngredients(recipeId: number) {
+        const recipe = await Recipe.query()
+            .where('id', recipeId)
+            .preload('ingredients', (query) => query.pivotColumns(['quantity', 'unit']))
+            .firstOrFail()
+        
+        const ingredients: RecipeIngredientDTO[] = recipe.ingredients.map((ingredient) => ({
+            id: ingredient.id,
+            recipeId: recipe.id,
+            ingredientId: ingredient.id,
+            quantity: ingredient.$extras.quantity,
+            unit: ingredient.$extras.unit
+        }))
 
-        return recipe.ingredients
+        return ingredients
+    }
+    
+    public async showRecipeIngredient(recipeId: number, ingredientId: number) {
+        const recipe = await Recipe.query().where('id', recipeId).preload('ingredients', (query) => {
+            query.where('ingredients.id', ingredientId)
+        })
+        .firstOrFail()
+
+        const ingredient = recipe.ingredients[0]
+
+        if (!ingredient) {
+            throw new Error('Ingredient not found in this recipe')
+        }
+
+        return {
+            id: ingredient.id,
+            name: ingredient.name,
+            quantity: ingredient.$extras.quantity,
+            unit: ingredient.$extras.unit,
+        }
     }
 }
